pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dailystream"
version = "0.1.0"
description = "A minimal daily recording stream system for macOS"
requires-python = ">=3.11"
dependencies = [
    "rumps>=0.4.0",
    "click>=8.1",
    "pyobjc-framework-Quartz>=10.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
dailystream = "dailystream.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/dailystream"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
